+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   `uptime` and `memory usage`.
 * Added rendered Midway/Taito copyright.
 * Refactored the event handling.
+* Improved SDL IO controller audio rendering.
 
 0.6.1 [04/08/24]
 * Added profiles for improved build support.